Database: ActiveRow refactored to Row
