+++ Nette/Database/Table/ActiveRow.php
@@ -22,13 +22,8 @@ use Nette,
  * @author     Jakub Vrana
  * @author     Jan Skrasek
  */
-class ActiveRow implements \IteratorAggregate, IRow
+class ActiveRow extends Row
 {
-	/** @var Selection */
-	private $table;
-
-	/** @var array of row data */
-	private $data;
 
 	/** @var bool */
 	private $dataRefreshed = FALSE;
@@ -37,39 +32,17 @@ class ActiveRow implements \IteratorAggregate, IRow
 	private $isModified = FALSE;
 
 
-	public function __construct(array $data, Selection $table)
-	{
-		$this->data = $data;
-		$this->table = $table;
-	}
-
 
 	/**
 	 * @internal
 	 * @ignore
 	 */
-	public function setTable(Selection $table)
+	public function setTable($selection)
 	{
-		$this->table = $table;
-	}
-
-
-	/**
-	 * @internal
-	 */
-	public function getTable()
-	{
-		return $this->table;
-	}
-
-
-	public function __toString()
-	{
-		try {
-			return (string) $this->getPrimary();
-		} catch (\Exception $e) {
-			trigger_error("Exception in " . __METHOD__ . "(): {$e->getMessage()} in {$e->getFile()}:{$e->getLine()}", E_USER_ERROR);
+		if (!$selection instanceof Selection) {
+			throw new \LogicException('Row table must be instance of Nette\Database\Table\Selection.');
 		}
+		$this->table = $selection;
 	}
 
 
@@ -79,89 +52,7 @@ class ActiveRow implements \IteratorAggregate, IRow
 	public function toArray()
 	{
 		$this->accessColumn(NULL);
-		return $this->data;
-	}
-
-
-	/**
-	 * Returns primary key value.
-	 * @param  bool
-	 * @return mixed possible int, string, array, object (Nette\DateTime)
-	 */
-	public function getPrimary($need = TRUE)
-	{
-		$primary = $this->table->getPrimary($need);
-		if ($primary === NULL) {
-			return NULL;
-
-		} elseif (!is_array($primary)) {
-			if (isset($this->data[$primary])) {
-				return $this->data[$primary];
-			} elseif ($need) {
-				throw new Nette\InvalidStateException("Row does not contain primary $primary column data.");
-			} else {
-				return NULL;
-			}
-
-		} else {
-			$primaryVal = array();
-			foreach ($primary as $key) {
-				if (!isset($this->data[$key])) {
-					if ($need) {
-						throw new Nette\InvalidStateException("Row does not contain primary $key column data.");
-					} else {
-						return NULL;
-					}
-				}
-				$primaryVal[$key] = $this->data[$key];
-			}
-			return $primaryVal;
-		}
-	}
-
-
-	/**
-	 * Returns row signature (composition of primary keys)
-	 * @param  bool
-	 * @return string
-	 */
-	public function getSignature($need = TRUE)
-	{
-		return implode('|', (array) $this->getPrimary($need));
-	}
-
-
-	/**
-	 * Returns referenced row.
-	 * @param  string
-	 * @param  string
-	 * @return IRow or NULL if the row does not exist
-	 */
-	public function ref($key, $throughColumn = NULL)
-	{
-		if (!$throughColumn) {
-			list($key, $throughColumn) = $this->table->getDatabaseReflection()->getBelongsToReference($this->table->getName(), $key);
-		}
-
-		return $this->getReference($key, $throughColumn);
-	}
-
-
-	/**
-	 * Returns referencing rows.
-	 * @param  string
-	 * @param  string
-	 * @return GroupedSelection
-	 */
-	public function related($key, $throughColumn = NULL)
-	{
-		if (strpos($key, '.') !== FALSE) {
-			list($key, $throughColumn) = explode('.', $key);
-		} elseif (!$throughColumn) {
-			list($key, $throughColumn) = $this->table->getDatabaseReflection()->getHasManyReference($this->table->getName(), $key);
-		}
-
-		return $this->table->getReferencingTable($key, $throughColumn, $this[$this->table->getPrimary()]);
+		return parent::toArray();
 	}
 
 
@@ -213,56 +104,12 @@ class ActiveRow implements \IteratorAggregate, IRow
 	public function getIterator()
 	{
 		$this->accessColumn(NULL);
-		return new \ArrayIterator($this->data);
+		return parent::getIterator();
 	}
 
 
 	/********************* interface ArrayAccess & magic accessors ****************d*g**/
 
-
-	/**
-	 * Stores value in column.
-	 * @param  string column name
-	 * @param  string value
-	 * @return void
-	 */
-	public function offsetSet($key, $value)
-	{
-		$this->__set($key, $value);
-	}
-
-
-	/**
-	 * Returns value of column.
-	 * @param  string column name
-	 * @return string
-	 */
-	public function offsetGet($key)
-	{
-		return $this->__get($key);
-	}
-
-
-	/**
-	 * Tests if column exists.
-	 * @param  string column name
-	 * @return bool
-	 */
-	public function offsetExists($key)
-	{
-		return $this->__isset($key);
-	}
-
-
-	/**
-	 * Removes column from data.
-	 * @param  string column name
-	 * @return void
-	 */
-	public function offsetUnset($key)
-	{
-		$this->__unset($key);
-	}
 
 
 	public function __set($key, $value)
@@ -303,11 +150,6 @@ class ActiveRow implements \IteratorAggregate, IRow
 	}
 
 
-	public function __unset($key)
-	{
-		throw new Nette\DeprecatedException('ActiveRow is read-only.');
-	}
-
 
 	protected function accessColumn($key, $selectColumn = TRUE)
 	{
@@ -328,13 +170,7 @@ class ActiveRow implements \IteratorAggregate, IRow
 	protected function getReference($table, $column)
 	{
 		$this->accessColumn($column);
-		if (array_key_exists($column, $this->data)) {
-			$value = $this->data[$column];
-			$referenced = $this->table->getReferencedTable($table, $column, $value);
-			return isset($referenced[$value]) ? $referenced[$value] : NULL; // referenced row may not exist
-		}
-
-		return FALSE;
+		return parent::getReference($table, $column);
 	}
 
 }
