+++ Nette/Database/Table/IRow.php
@@ -22,7 +22,7 @@ use Nette\Database;
 interface IRow extends Database\IRow
 {
 
-	function setTable(Selection $name);
+	function setTable($resultSet);
 
 	function getTable();
